kata 2.0

+++ src/main/java/ru/kata/spring/boot_security/demo/model/User.java
@@ -9,129 +9,6 @@ import javax.persistence.*;
 import java.util.Collection;
 import java.util.Set;
 
-//@Entity
-//@Table(name = "users")
-//public class User implements UserDetails {
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//    @Column
-//    private String name;
-//    @Column
-//    private String surname;
-//    @Column
-//    private int age;
-//    @Column
-//    private String email;
-//    @Column
-//    private String password;
-//
-//    @ManyToMany(fetch = FetchType.EAGER)
-//    @JoinTable(name = "user_role",
-//            joinColumns = @JoinColumn(name = "user_id"),
-//            inverseJoinColumns = @JoinColumn(name = "role_id"))
-//    private Set<Role> roles;
-//
-//
-//    public User() {
-//
-//    }
-//
-//    public User(String email, String password, Collection<? extends GrantedAuthority> authorities) {
-//    }
-//
-//
-//    public Long getId() {
-//        return id;
-//    }
-//
-//    public void setId(Long id) {
-//        this.id = id;
-//    }
-//
-//    public String getName() {
-//        return name;
-//    }
-//
-//    public void setName(String name) {
-//        this.name = name;
-//    }
-//
-//    public String getSurname() {
-//        return surname;
-//    }
-//
-//    public void setSurname(String surname) {
-//        this.surname = surname;
-//    }
-//
-//    public int getAge() {
-//        return age;
-//    }
-//
-//    public void setAge(int age) {
-//        this.age = age;
-//    }
-//
-//    public String getEmail() {
-//        return email;
-//    }
-//
-//    public void setEmail(String email) {
-//        this.email = email;
-//    }
-//
-//    public void setPassword(String password) {
-//        this.password = password;
-//    }
-//
-//
-//    public Set<Role> getRole() {
-//        return roles;
-//    }
-//
-//    public void setRole(Set<Role> role) {
-//        this.roles = role;
-//    }
-//
-//    @Override
-//    public Collection<? extends GrantedAuthority> getAuthorities() {
-//        return getRole();
-//    }
-//
-//    @Override
-//    public String getPassword() {
-//        return password;
-//    }
-//
-//    @Override
-//    public String getUsername() {
-//        return email;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonLocked() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isCredentialsNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isEnabled() {
-//        return true;
-//    }
-//
-//}
-
 @Data
 @Entity
 @Table(name = "users")
@@ -139,7 +16,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "name", unique = true)
     private String name;
@@ -164,7 +41,7 @@ public class User implements UserDetails {
     private Set<Role> roles;
 
 
-    //    @ManyToMany(fetch = FetchType.EAGER)
+//        @ManyToMany(fetch = FetchType.EAGER)
 //    @JoinTable(name = "user_role",
 //            joinColumns = @JoinColumn(name = "user_id"),
 //            inverseJoinColumns = @JoinColumn(name = "role_id"))
@@ -177,7 +54,7 @@ public class User implements UserDetails {
         return getRoles();
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
